Implement optimistic deletion of courses
* Add relevant action and reducer
* Wire them up in CoursesPage and CourseList components

Optimistic updates is an interesting concept meant to make UX fast and
responsive. In a standard async interaction, we call an API, wait for
it to return successfully, and then update our UI accordingly. However,
in the optimistic approach we update UI as soon as our async call is
initiated. We DO NOT wait for a successful response in order to make
things faster and snappier for the user. If an error does occur, we
handle it in catch() and show a relevant toast.