Add React Testing Library tests for CourseForm
We rewrite the same unit tests as we did for this component in Enzyme.
This is to contrast how both libraries are different.

Enzyme lets you test a component in isolation (shallow) or
as it is rendered in DOM (mount), and provides a jQuery-like query
system to get values.

React Testing Library, on the other hand, is based on the philosophy of
testing components as users see them (that is, fully rendered in DOM).
For that reason, rendered components are always 'mounted' in RTL.
Also, instead of providing a selector-based query system to get values,
RTL provides ByText, ByPlaceholder, ByTitle, ByAltText, etc.
query functions to find text as users see it in browser (rather than
inside a DOM element).